fix: user profile modal
